@@ -16,7 +16,7 @@
     class="hidden absolute -top-[calc(42%-3px)] left-[calc(42%-3px)] w-1 h-36 -rotate-45 group-hover/github-corner:opacity-75 group-focus-visible/github-corner:block pointer-events-none"
     class:bg-white={primaryBackground}
     class:bg-neutral-900={!primaryBackground}
-  />
+  ></div>
   <svg
     aria-hidden="true"
     class:fill-white={primaryBackground}
