@@ -1,4 +1,4 @@
-<script>
+<script lang="ts">
   import { page } from '$app/stores';
   import Card from '$lib/components/Card.svelte';
 </script>
