@@ -1,4 +1,4 @@
-import type { Project } from '$lib/types'
+import type { Project } from '$lib/types';
 
 /** @type {import('./$types').PageLoad} */
 export async function load({ fetch }) {
