@@ -1,9 +1,5 @@
-import { browser } from '$app/environment';
+const webmanifest = document.getElementById('webmanifest') as HTMLLinkElement;
 
-if (browser) {
-  const webmanifest = document.querySelector('link[rel="manifest"]') as HTMLLinkElement;
-
-  if (navigator.userAgent.search('Mac') !== -1 && webmanifest) {
-    webmanifest.href = '/manifest-apple.webmanifest';
-  }
+if (import.meta.env.PROD && navigator.userAgent.search('Mac') !== -1 && webmanifest) {
+  webmanifest.href = '/manifest-apple.webmanifest';
 }
