@@ -1,4 +1,5 @@
-/** @type {import('tailwindcss').Config} */
+import type { Config } from 'tailwindcss';
+
 export default {
   content: ['./src/**/*.{html,js,svelte,ts}'],
   theme: {
@@ -41,9 +42,9 @@ export default {
           '40%, 80%': {
             transform: 'rotate(10deg)',
           },
-        }
-      }
+        },
+      },
     },
   },
   plugins: [],
-}
+} satisfies Config;
